refactor structure of sb
